Componentizes Typer
